+++ db/migrate/20220423222023_create_thirds.rb
@@ -2,7 +2,7 @@ class CreateThirds < ActiveRecord::Migration[6.1]
   def change
     create_table :thirds do |t|
       t.string :title
-      t.integer :level
+      t.float :level
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
